+++ allegro_shim/include/allegro5/allegro_audio.h
@@ -101,6 +101,7 @@ ALLEGRO_MIXER* al_create_mixer(unsigned int freq, ALLEGRO_AUDIO_DEPTH depth, ALL
 void al_destroy_mixer(ALLEGRO_MIXER* mixer);
 bool al_attach_sample_instance_to_mixer(ALLEGRO_SAMPLE_INSTANCE* stream, ALLEGRO_MIXER* mixer);
 bool al_mixer_attach_sample(ALLEGRO_MIXER* mixer, ALLEGRO_SAMPLE* sample);
+bool al_mixer_detach_sample(ALLEGRO_MIXER* mixer);
 bool al_attach_audio_stream_to_mixer(ALLEGRO_AUDIO_STREAM* stream, ALLEGRO_MIXER* mixer);
 unsigned int al_get_mixer_frequency(const ALLEGRO_MIXER* mixer);
 ALLEGRO_CHANNEL_CONF al_get_mixer_channels(const ALLEGRO_MIXER* mixer);